Refactor Integration Tests
Examining response for proper object structure happens a lot, so I defined this JSON structure within karate-config.js in order to reuse throughout the features.

Additionally, the way that the URL global variable was accessed was unnecessary complex. I removed it, and in doing so no longer had a reason to keep the 'Background' section. Tests are still ordered by the csv and arranged by Controllers, which could probably be improved.