Correcting Merge, Another bump to read time
When I merged, I did not accurately retain the selection of all Controllers. This commit fixes that.

I bumped up the read time from 10s to 15s after seeing a failure. I profiled the call with Chrome dev tools and see that it's loading at just over 10s, so 15s should be the sweet spot for this.