Ajax call and local readme
Jenkinsfile takes care of grabbing the readme, converting it into a text file, and depositing it in /static folder so it can be retrieved.

Ajax call replaces the synchronous http GET request for the raw readme data (which is now pointed locally). The rest of the delivery mechanism remains the same.